requirements.txt
# requirements.txt
# Core dependencies
tweepy>=4.14.0
requests==2.28.1
beautifulsoup4==4.11.1
pandas==2.2.2
openai>=1.0.0
python-dotenv==0.21.0

# Financial analysis
finnhub-python>=2.4.18
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
ta>=0.10.0

# Testing and development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Logging and monitoring
python-json-logger>=2.0.7
structlog>=23.1.0
psutil>=5.9.0

# Scheduling
apscheduler>=3.10.1

# System monitoring
prometheus-client>=0.17.0

# Google Cloud
google-cloud-logging>=3.5.0
google-cloud-monitoring>=2.13.0